adapter: extract begin_statement_logging helper from frontend_peek
The new statement and reused-outer-context branches in
`try_frontend_peek` are otherwise identical bookkeeping. Lift them
into a `PeekClient::begin_statement_logging` method so that another
frontend-sequenced caller (the upcoming OCC read-then-write path)
can share the same logic.

Work towards https://github.com/MaterializeInc/database-issues/issues/6686
